refer to example
If the last part of this section simply comments that the nature of the spiral can be determined by considering the eigenvectors, the reader may panic and assume that they are supposed to understand immediately how to do this. We must point them to the example given in the next section.