implement LuaTeX support
When the luatex backend of hyperref is used with default options, some definitions (macros) are set* which conflict with our maths symbol definitions. In consequence, notes in this repository cannot be compiled with LuaLaTeX.

This commit implements a hack whereby these conflicting definitions are unset before we need to use them. This hack is not very elegant. A better solution would involve namespaces, which unfortunately LateX doesn't provide.

* See the "pdfencoding" option in the hyperref docs.